refactor: Add with_middlewares and with_cache methods to TestFetchClient
This commit adds two new methods, `with_middlewares` and `with_cache`, to the `TestFetchClient` struct in the `fetch` module. These methods allow for configuring middlewares and cache for the client. The `with_middlewares` method takes a vector of middlewares and sets them on the client, while the `with_cache` method takes a `Rc<RefCell<dyn Cacheable>>` and sets it as the cache for the client.

The changes were made in the `fetch/src/lib.rs` file.